random tweaks

+++ tools/quake2/q2map/trace.c
@@ -87,7 +87,7 @@ void MakeTnode( int nodenum ){
 void MakeTnodes( dmodel_t *bm ){
 	// 32 byte align the structs
 	tnodes = malloc( ( numnodes + 1 ) * sizeof( tnode_t ) );
-	tnodes = (tnode_t *)( ( (int)tnodes + 31 ) & ~31 );
+	tnodes = (tnode_t *)( ( (size_t)tnodes + 31 ) & ~31 );
 	tnode_p = tnodes;
 
 	MakeTnode( 0 );
